Fix C4267, add NiSyncVector,
NiSyncVector can be used for structs/class members that require custom IO (sync functions).

+++ include/KDMatcher.hpp
@@ -276,7 +276,7 @@ public:
 		root->find_closest(querypoint, queryResult, radius, mindist);
 		std::sort(queryResult.begin(), queryResult.end());
 
-		return queryResult.size();
+		return static_cast<int>(queryResult.size());
 	}
 };
 } // namespace nifly
